Precompute auction like solvable orders
isntead of creating it on demand.
