+++ crates/orderbook/src/api/get_solvable_orders.rs
@@ -13,7 +13,7 @@ pub fn get_solvable_orders(
     get_solvable_orders_request().and_then(move || {
         let orderbook = orderbook.clone();
         async move {
-            let result = orderbook.get_solvable_orders().await;
+            let result = orderbook.get_solvable_orders();
             Result::<_, Infallible>::Ok(convert_json_response(
                 result.map(|solvable_orders| solvable_orders.orders),
             ))
